m3placement: Moved to own module under common
This is mainly so that we can use it from elsewhere, and the
functionality is relatively isolated. The code itself should be only
used from coordinator's m3db plugin within astacus codebase.
